eOx5BZ3C: created aggregated_event
