+++ app/models/publish_services_metadata_event.rb
@@ -1,0 +1,13 @@
+
+class PublishServicesMetadataEvent < Event
+  data_attributes :cert_config
+  before_create :populate_data_attributes
+
+  def populate_data_attributes
+    self.assign_attributes(attributes_to_apply)
+  end
+
+  def attributes_to_apply
+    {cert_config: "I am here" }
+  end
+end
