+++ app/models/upload_certificate_event.rb
@@ -1,4 +1,4 @@
-class UploadCertificateEvent < Event
+class UploadCertificateEvent < AggregatedEvent
   belongs_to_aggregate :certificate
   data_attributes :value, :usage, :component_id
   before_save :convert_value_to_inline_der
