+++ spec/models/publish_services_event_metadata_spec.rb
@@ -1,0 +1,10 @@
+require 'rails_helper'
+
+RSpec.describe PublishServicesMetadataEvent, type: :model do
+  context '#create' do
+    it 'creates an event which contains hard-coded data' do
+      event = PublishServicesMetadataEvent.create()
+      expect(event.data).to include("cert_config" => "I am here" )
+    end
+  end
+end
